[Improve] Move HomeDrawerChecklistItemModel to domain layer
