+++ app/src/main/java/wottrich/github/io/smartchecklist/domain/model/HomeDrawerChecklistItemModel.kt
@@ -1,4 +1,4 @@
-package wottrich.github.io.smartchecklist.presentation.ui.model
+package wottrich.github.io.smartchecklist.domain.model
 
 data class HomeDrawerChecklistItemModel(
     val checklistUuid: String,
