+++ app/src/main/java/wottrich/github/io/smartchecklist/domain/usecase/GetChecklistDrawerUseCase.kt
@@ -8,7 +8,7 @@ import wottrich.github.io.smartchecklist.coroutines.FlowableUseCase
 import wottrich.github.io.smartchecklist.coroutines.UseCase
 import wottrich.github.io.smartchecklist.coroutines.base.Result
 import wottrich.github.io.smartchecklist.domain.mapper.HomeDrawerChecklistItemModelMapper
-import wottrich.github.io.smartchecklist.presentation.ui.model.HomeDrawerChecklistItemModel
+import wottrich.github.io.smartchecklist.domain.model.HomeDrawerChecklistItemModel
 
 class GetChecklistDrawerUseCase(
     private val checklistRepository: ChecklistRepository,
